// src/shared/enums/CharacterState.ts
export enum CharacterState {
    UNKNOWN = '',
    IDLE = 'IDLE',
    IN_AIR = 'IN_AIR',
    ON_GROUND = 'ON_GROUND',
    START_JUMP = 'START_JUMP',
    JUMPING = 'JUMPING',
    FALLING = 'FALLING',
    CLOSE_TO_LAND = 'CLOSE_TO_LAND',
    FALLING_TO_CRASH = 'FALLING_TO_CRASH',
    CLOSE_TO_CRASH = 'CLOSE_TO_CRASH',
    WAS_SHOOT = 'WAS_SHOOT',
    IS_DYING = 'IS_DYING',
    RUNNING = 'RUNNING',
    WALKING_BACKWARDS = 'WALKING_BACKWARDS',
    HAS_WON = 'HAS_WON',
    THROWING_FREESBE_GROUND = 'THROWING_FREESBE_GROUND'
}