// src/shared/enums/CharacterIntention.ts
export enum CharacterIntention {

    DO_NOTHING = 'DO_NOTHING',

    WANTS_TO_GO_FORWARD = 'WANTS_TO_GO_FORWARD',
    WANTS_TO_GO_BACKWARDS = 'WANTS_TO_GO_BACKWARDS',
    WANTS_TO_TURN_RIGHT = 'WANTS_TO_TURN_RIGHT',
    WANTS_TO_TURN_LEFT = 'WANTS_TO_TURN_LEFT',

    WANTS_TO_JUMP = 'WANTS_TO_JUMP',
    WANTS_TO_THROW_FREESBE = 'WANTS_TO_THROW_FREESBE',
    WANTS_TO_ROLL = 'WANTS_TO_ROLL',
    WANTS_TO_RUN = 'WANTS_TO_RUN',
    WANTS_TO_DUCK = 'WANTS_TO_DUCK',

    
}